Move Mocks to testdata folder
