+++ testdata/mocks/mock_resource.go
